Downgrade KDE runtime to 5.15-22.08
If I remember correctly, the only real reason we were using the KDE 6
runtime in the first place, was because of a bad file descriptor leak
that would happen whenever we hovered the mouse cursor over any
clickable URLs in the article viewer panel.

This issue can still be reproduced with the official AppImage, which
still bundles Qt 5.15.4 from Ubuntu 20.04.

However, because the KDE runtime 5.15 for Flatpak uses a more recent
version of Qt 5, this particular issue has been fixed already.